+++ lua-api-crates/mux/src/tab.rs
@@ -124,5 +124,31 @@ impl UserData for MuxTab {
             let tab = this.resolve(&mux)?;
             to_lua(lua, tab.get_size())
         });
+
+        methods.add_method("activate", move |_lua, this, ()| {
+            let mux = Mux::get();
+            let tab = this.resolve(&mux)?;
+
+            let pane = tab.get_active_pane().ok_or_else(|| {
+                mlua::Error::external(format!("tab {} has no active pane!?", this.0))
+            })?;
+
+            let (_domain_id, window_id, tab_id) =
+                mux.resolve_pane_id(pane.pane_id()).ok_or_else(|| {
+                    mlua::Error::external(format!("pane {} not found", pane.pane_id()))
+                })?;
+            {
+                let mut window = mux.get_window_mut(window_id).ok_or_else(|| {
+                    mlua::Error::external(format!("window {window_id} not found"))
+                })?;
+                let tab_idx = window.idx_by_id(tab_id).ok_or_else(|| {
+                    mlua::Error::external(format!(
+                        "tab {tab_id} isn't really in window {window_id}!?"
+                    ))
+                })?;
+                window.save_and_then_set_active(tab_idx);
+            }
+            Ok(())
+        });
     }
 }